meta-qti-bt: enable SOM configuration for Naples and add audiohal dependency
Remove redundant denpendencies from bluetooth recipes, and add system-core-
headers and media-headers to recipes.
Append to OE configuration for Naples in fluoride and libbt-vendor recipes
to enable SOM configuration when naples feature is enabled in MACHINE_FEATURES.
Rome, Pronto and Cherokee also follow the change of OE configuration.
Qti-audio in DISTRO_FEATURES is used to decide if audio dependency is needed
in bt-app and fluoride recipes.

Change-Id: Ia064cd9424fdece61eed2c52aedbeb64d1a9d2bf
Signed-off-by: Liuliu Zhao <liulzhao@codeaurora.org>